change id of entity  from long to int re infrastructure again web project

+++ Source/router-service/artifacter/src/main/java/com/fpt/router/artifacter/model/algorithm/PathInfo.java
@@ -16,7 +16,7 @@ public class PathInfo implements IAlgorithmModel {
 
     }
 
-    public long pathInfoId;
+    public int pathInfoId;
 
     public int pathInfoNo;
